Remove duplicated files and moved group tests to custom tests dir
