+++ _tests/custom_tests/tests/RunGroupTest.php
@@ -1,11 +1,9 @@
 <?php
 
-use QIT\SelfTests\GroupTests\Traits\SnapshotHelpers;
-use Spatie\Snapshots\MatchesSnapshots;
+use QIT\SelfTests\CustomTests\Traits\SnapshotHelpers;
 
 class RunGroupTest extends \PHPUnit\Framework\TestCase {
 	use SnapshotHelpers;
-	use MatchesSnapshots;
 
 	public function test_run_remote_group() {
 		// Clear any existing group.
@@ -61,7 +59,7 @@ class RunGroupTest extends \PHPUnit\Framework\TestCase {
 
 
 		$this->assertStringContainsString( 'The test group is already registered. Please execute the group with `group:run` and delete it with `group:clear` after it has completed.', $output );
-		
+
 		// Run Group.
 		$output = qit( [
 			'group:run'
@@ -116,7 +114,7 @@ class RunGroupTest extends \PHPUnit\Framework\TestCase {
 		$output = qit( [
 			'group:clear',
 		] );
-		
+
 		$tests = [
             [
                 'type'        => 'security',
@@ -274,16 +272,16 @@ class RunGroupTest extends \PHPUnit\Framework\TestCase {
 			'group:run',
 		] );
 
-		
+
 		// At this point, both local tests would be done and we can retrieve their results from remote.
 		$output = qit( [
 			'group:show',
 		] );
 
-		
+
 		$normalized_output = $this->normalize_complete_group_run_output( $output, true );
 		$this->assertMatchesSnapshot( $normalized_output );	
-		
+
 	}
 
 	public function test_local_group_run_with_remote_and_local_tests() {
@@ -291,7 +289,7 @@ class RunGroupTest extends \PHPUnit\Framework\TestCase {
 		$output = qit( [
 			'group:clear',
 		] );
-		
+
 		$tests = [
             [
                 'type'        => 'activaiton',
@@ -329,13 +327,13 @@ class RunGroupTest extends \PHPUnit\Framework\TestCase {
 			'group:run',
 		] );
 
-		
+
 		// At this point, both local tests would be done and we can retrieve their results from remote.
 		$output = qit( [
 			'group:show',
 		] );
 
-		
+
 		$normalized_output = $this->normalize_complete_group_run_output( $output, true );
 		$this->assertMatchesSnapshot( $normalized_output );	
 	}
