--- _tests/group_tests/composer.json
@@ -1,27 +1,0 @@
-{
-    "config": {
-      "platform": {
-        "php": "8.2"
-      }
-    },
-    "require": {
-      "php": "^8"
-    },
-    "autoload": {
-      "psr-4": {
-        "QIT\\SelfTests\\GroupTests\\": "tests"
-      },
-      "files": [
-        "QITTestExtension.php"
-      ]
-    },
-    "require-dev": {
-      "phpunit/phpunit": "^11",
-      "spatie/phpunit-snapshot-assertions": "^5.0",
-      "symfony/process": "^7",
-      "symfony/filesystem": "^7",
-      "vlucas/phpdotenv": "^5",
-      "brianium/paratest": "^7"
-    }
-  }
-
